Factoring out strutils from amcl_node
